@@ -11,30 +11,31 @@
   },
   "main": "server.js",
   "keywords": [
-    "universal",
-    "isomorphic",
+    "Blog",
+    "vantu19000",
     "react",
     "tutorial"
   ],
   "dependencies": {
     "axios": "^0.18.0",
-    "babel-core": "^6.9.0",
-    "babel-loader": "^6.2.4",
-    "babel-preset-es2015": "^6.24.0",
-    "babel-preset-es2017": "^6.24.1",
-    "babel-preset-react": "^6.5.0",
-    "babel-register": "^6.9.0",
-    "express": "^4.13.4",
-    "react": "^15.0.2",
-    "react-dom": "^15.0.2",
-    "react-redux": "^4.4.10",
-    "react-router": "^2.8.1",
-    "redux": "^3.7.2",
-    "webpack": "^1.13.1"
+    "babel-core": "^6.26.3",
+    "babel-loader": "^8.0.6",
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-react": "^6.24.1",
+    "babel-register": "^6.26.0",
+    "express": "^4.16.4",
+    "npm": "^6.9.0",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
+    "react-redux": "^7.0.3",
+    "react-router": "^5.0.0",
+    "redux": "^4.0.1",
+    "redux-thunk": "^2.3.0",
+    "webpack": "^4.31.0"
   },
   "devDependencies": {
     "nodemon": "^1.19.0",
-    "npm-run-all": "^3.1.2",
+    "npm-run-all": "^4.1.5",
     "webpack-cli": "^3.3.2"
   },
   "private": true,
